Token e tipoUsuário para localStorage.
